refactor: decouple PlaceItems from PlaceMaps
